requirements.txt
asgiref==3.3.4
certifi==2020.12.5
click==8.0.1
Django==3.2.3
django-environ==0.4.5
djangorestframework==3.12.4
elasticsearch==7.13.0
pytz==2021.1
redis==3.5.3
rq==1.8.1
sqlparse==0.4.1
urllib3==1.26.5
